+++ CHANGES.md
@@ -29,6 +29,7 @@
 * Make policy at update_num=0 sampled from Domain's sample_state_action
 * Separate timings for HER and rb
 * Video added to viz when visualizing solutions
+* Can add random edges to graph and beam search when using policy
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search